feat: Change output mode to 'hybrid'
This commit changes the output mode of the project to 'hybrid', meaning that only the 'index' page will be rendered on demand and the rest of the pages will be pre-rendered as static pages. This allows the 'index' page to change its content whenever the YouTube API receives new data from the YouTube channel, when uploading a new video. This is reflected by setting the 'prerender' constant to 'false' inside the 'index' page.

Some text changes are also introduced in the News component. The netlify integration is also installed with this commit.